Double precision timeoutz is betterz

+++ paxos/BallotNumber.java
@@ -3,9 +3,9 @@ package paxos;
 public class BallotNumber implements Comparable {
 	int round;
 	ProcessId leader_id;
-	long timeout;
+	double timeout;
 
-	public BallotNumber(int round, ProcessId leader_id, long timeout){
+	public BallotNumber(int round, ProcessId leader_id, double timeout){
 		this.round = round;
 		this.leader_id = leader_id;
 		this.timeout = timeout;
